refactor: create separate controller for articles by search
