+++ app/controllers/stories_controller.rb
@@ -16,8 +16,8 @@ class StoriesController < ApplicationController
 
   SIGNED_OUT_RECORD_COUNT = 60
 
-  before_action :authenticate_user!, except: %i[index search show]
-  before_action :set_cache_control_headers, only: %i[index search show]
+  before_action :authenticate_user!, except: %i[index show]
+  before_action :set_cache_control_headers, only: %i[index show]
   before_action :redirect_to_lowercase_username, only: %i[index]
 
   rescue_from ArgumentError, with: :bad_request
@@ -29,14 +29,6 @@ class StoriesController < ApplicationController
     return handle_user_or_organization_or_podcast_or_page_index if params[:username]
 
     handle_base_index
-  end
-
-  def search
-    @query = "...searching"
-    @article_index = true
-    @current_ordering = current_search_results_ordering
-    set_surrogate_key_header "articles-page-with-query"
-    render template: "articles/search"
   end
 
   def show
@@ -441,11 +433,5 @@ class StoriesController < ApplicationController
       @user.github_username.present? ? "https://github.com/#{@user.github_username}" : nil,
       @user.website_url,
     ].reject(&:blank?)
-  end
-
-  def current_search_results_ordering
-    return :relevance unless params[:sort_by] == "published_at" && params[:sort_direction].present?
-
-    params[:sort_direction] == "desc" ? :newest : :oldest
   end
 end
