+++ app/views/stories/articles_search/_nav_menu.html.erb
