+++ app/views/stories/articles_search/index.html.erb
@@ -1,5 +1,5 @@
 <%= content_for :page_meta do %>
-  <%= render "articles/search/meta" %>
+  <%= render "stories/articles_search/meta" %>
 <% end %>
 <div id="query-wrapper"></div>
 
@@ -38,7 +38,7 @@
 
     <div id="sidebar-wrapper-left" class="sidebar-wrapper sidebar-wrapper-left">
       <div class="sidebar-bg" id="sidebar-bg-left"></div>
-      <%= render "articles/search/nav_menu" %>
+      <%= render "stories/articles_search/nav_menu" %>
     </div>
 
     <div class="articles-list crayons-layout__content" id="articles-list" data-follow-button-container="true">
